+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 * Added component(1) support ([view](https://github.com/videojs/video.js/pull/1032))
 * Captions now move down when controls are hidden ([view](https://github.com/videojs/video.js/pull/1053))
 * Added the .less source file to the distribution files ([view](https://github.com/videojs/video.js/pull/1056))
+* Changed src() to return the current selected source ([view](https://github.com/videojs/video.js/pull/968))
 
 --------------------